Project/backend/requirements.txt
Flask
flask-cors
python-dotenv
pandas
numpy
pillow
torch
joblib
xgboost
